Move fuel config to DvvModificationsServiceClient
