+++ service/src/main/kotlin/fi/espoo/evaka/dvv/DvvModificationsServiceClient.kt
@@ -13,7 +13,17 @@ import com.github.kittinunf.fuel.core.Request
 import com.github.kittinunf.fuel.core.extensions.jsonBody
 import com.github.kittinunf.result.Result
 import fi.espoo.evaka.DvvModificationsEnv
+import fi.espoo.evaka.EvakaEnv
+import fi.espoo.evaka.KeystoreEnv
+import fi.espoo.evaka.VtjXroadEnv
+import java.security.cert.X509Certificate
 import java.time.LocalDate
+import javax.net.ssl.HostnameVerifier
+import javax.net.ssl.KeyManagerFactory
+import javax.net.ssl.SSLContext
+import javax.net.ssl.TrustManager
+import javax.net.ssl.TrustManagerFactory
+import javax.net.ssl.X509TrustManager
 import mu.KotlinLogging
 import org.springframework.stereotype.Service
 
@@ -26,14 +36,81 @@ private val logger = KotlinLogging.logger {}
 @Service
 class DvvModificationsServiceClient(
     private val jsonMapper: JsonMapper,
-    private val fuel: FuelManager,
     private val customizers: List<DvvModificationRequestCustomizer>,
+    evakaEnv: EvakaEnv,
+    xroadEnv: VtjXroadEnv,
     env: DvvModificationsEnv
 ) {
     private val serviceUrl: String = env.url
     private val dvvUserId = env.userId
     private val dvvPassword = env.password
     private val dvvXroadClientId: String = env.xroadClientId
+
+    private val fuel =
+        if (
+            evakaEnv.httpClientCertificateCheck &&
+                xroadEnv.keyStore != null &&
+                xroadEnv.trustStore != null
+        ) {
+            certCheckFuelManager(keyStore = xroadEnv.keyStore, trustStore = xroadEnv.trustStore)
+        } else if (!evakaEnv.httpClientCertificateCheck) {
+            noCertCheckFuelManager()
+        } else {
+            FuelManager()
+        }
+
+    private fun certCheckFuelManager(keyStore: KeystoreEnv, trustStore: KeystoreEnv): FuelManager =
+        FuelManager().apply {
+            val keyManagerFactory =
+                KeyManagerFactory.getInstance(KeyManagerFactory.getDefaultAlgorithm())
+            keystore = keyStore.load()
+            keyManagerFactory.init(keystore, keyStore.password?.value?.toCharArray())
+
+            val trustManagerFactory =
+                TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm())
+            val trustKeyStore = trustStore.load()
+            trustManagerFactory.init(trustKeyStore)
+
+            socketFactory =
+                SSLContext.getInstance("SSL")
+                    .apply {
+                        init(
+                            keyManagerFactory.keyManagers,
+                            trustManagerFactory.trustManagers,
+                            java.security.SecureRandom()
+                        )
+                    }
+                    .socketFactory
+
+            hostnameVerifier = HostnameVerifier { _, _ -> true }
+        }
+
+    private fun noCertCheckFuelManager() =
+        FuelManager().apply {
+            val trustAllCerts =
+                arrayOf<TrustManager>(
+                    object : X509TrustManager {
+                        override fun getAcceptedIssuers(): Array<X509Certificate>? = null
+
+                        override fun checkClientTrusted(
+                            chain: Array<X509Certificate>,
+                            authType: String
+                        ) = Unit
+
+                        override fun checkServerTrusted(
+                            chain: Array<X509Certificate>,
+                            authType: String
+                        ) = Unit
+                    }
+                )
+
+            socketFactory =
+                SSLContext.getInstance("SSL")
+                    .apply { init(null, trustAllCerts, java.security.SecureRandom()) }
+                    .socketFactory
+
+            hostnameVerifier = HostnameVerifier { _, _ -> true }
+        }
 
     // Fetch the first modification token of the given date
     fun getFirstModificationToken(
